+++ include/x86.h
@@ -8,7 +8,7 @@ bool a2286_init(struct autoconfig_info *aci);
 bool a2386_init(struct autoconfig_info *aci);
 bool isa_expansion_init(struct autoconfig_info *aci);
 void x86_bridge_sync_change(void);
-void x86_update_sound(double);
+void x86_update_sound(float);
 void x86_mouse(int port, int x, int y, int z, int b);
 
 #define X86_STATE_INACTIVE 0
